+++ packages/core-elements/src/elements/form-field/index.ts
@@ -3,3 +3,4 @@ export * from './form-field-error'
 export * from './form-field-help'
 export * from './form-field-label'
 export * from './form-field'
+export * from './use-form-field-state'
