+++ packages/core-elements/src/elements/form-field/use-form-field-state.ts
@@ -1,0 +1,15 @@
+import { useId } from 'react'
+
+export function useFormFieldState() {
+  const inputId = useId()
+  const labelId = useId()
+  const descriptionId = useId()
+  const errorId = useId()
+
+  return {
+    inputId,
+    labelId,
+    descriptionId,
+    errorId,
+  }
+}
